Add base rendering code to the Scene class
The Scene follows a design pattern that attempts to hide any references
to the rendering code in inherited classes. An Array containing
GameObjects is created to provide a way to render the objects upon
calling the render(...) method on the Scene.

Upon calling the render method, an operation is performed on all of the
GameObjects with the following steps:
 - verify that the spriteName isn't blank and the image isn't null
 - determine the type of the image (Texture/TextureRegion)
 - render according to the properties defined in the GameObject