Fix NaN issue upon touch input
Calling Gdx.input.delta[A] immediately after touch will result in a
force of NaN on both directions from not having a previous touch
position. This issue is solved by replacing an 'if' with an 'else if' on
the isTouched expression to skip the check upon the first touch.

This also fixes a future issue where touching the screen after a recent
release of the finger will result in the player moving erratically from
re-using the previous touch location.